Update to Java 11, schema topic management, PMD update
Big update to facilitate further upgrading. The tools now depend on Java
11. This was needed for some small features, but will keep images in
line with future Java development.

Main new feature: it is now possible to back up the entire _schemas
topic and to replace it with a correct one if it has become corrupted.
THis change was made to avoid corruptions as we encounter in RADAR-base
once in a while.